requirements.txt
alembic==1.16.5
asgiref==3.10.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
Flask==3.1.2
flask-cors==6.0.1
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mercadopago==2.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.5
setuptools==80.9.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
